requirements.txt
numpy
scipy
pyransac3d
pynetworktables
